Cleanup and initial setup for Lounge infra
This includes a bunch of stuff so it's hard to go into all of it, but it
is essentially a sanity pass on removing all the things that won't be
necessary for the local infrastructure that is being setup, along with
adding configs for setting up a proxmox host and provisioning new LXC
containers with ansible (terraform proxmox module was not playing nice
for multiple reasons).

Currently has three parts working -- host machine with disk setup and
mergerfs + snapraid, sharedom LXC for setting up a SMB server to share
/mnt/storage and /mnt/downloads (might be adjusted in the future), and
lounge LXC to contain the main plex-related functionalities.

Lounge is currently only running dozzle, authelia and traefik, and
traefik is having hiccups with cert generation so that's a WIP.

To-do includes:
- Reviewing current unpriviliged LXC -> root user -> 100000 mapping and
  if it makes sense to keep up with;
- Figuring out DRI pass-through for hardware acceleration; requires
  kernel update for proxmox;
- Figuring out migration from current Windows infrastructure to the new
  one to avoid having to redo everything around Plex, Radarr, Sonarr,
  Bazarr, NZBGet and Prowlarr;
- Setting up Authelia as auth provider and service config for the *arrs;
- Setting up Deluge docker container for downloads;
- Adding remaining disk after content copy is finished;
- Setting up global caddy / nginx / traefik LXC as reverse proxy for
  internal services, limit direct external access to containers, plus
  SSL logic and all that fun stuff;
- Setting up Tailscale VPN in host;
- Moving VPN services to a new LXC, refactor architecture to match,
  eventually set up gitea for self-sufficience;
- Maybe set up pihole;
- Maybe set up lightweight DHCP solution to avoid having to go back and
  forth to FGW;
- Setting up monitoring and alerts for server, proxmox containers and
  docker services;
- Figuring out extra services to set up (nextcloud, lychee, etc);